sb3: remove BlockData wrapper & utility functions
BlockData doesn't simplify enough code to be worth the inherent added
complexity of an unfamiliar data structure. Removing it simply makes
explicit the underlying behavior (i.e. Object.assign).

BlockData was more useful when it contained data besides simply the
blocks; originally, it was planned to contain both blocks and
broadcasts. It was a useful way to contain the data returned from a
variety of serialization functions. Now, only blocks are returned;
it's simpler to just refer to the actual sb3.Target["blocks"] type.